feat: 아카이브 채팅 조회 구현

+++ dbdeep-BE/src/main/java/kr/dbdeep/dbdeep_BE/domain/chat/dto/ChatMessageDto.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 public record ChatMessageDto(
         String id,
         String content,
-        Long memberId,
+        Integer memberId,
         String senderType,
         Timestamp timestamp
 ) {
